+++ backend/src/main.py
@@ -11,8 +11,8 @@ app = FastAPI()
 app.add_middleware(     
     CORSMiddleware,     
     allow_credentials=True,     
-    allow_origins=[""],     
-    allow_methods=[""],     
+    allow_origins=["*"],     
+    allow_methods=["*"],     
     allow_headers=["*"], 
 )
