+++ frontend/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --proxy-config proxy.conf.json --host 0.0.0.0 --disable-host-check true",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
