+++ frontend/src/app/app.module.ts
@@ -1,24 +1,22 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
-import { FormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { AppRoutingModule } from './app-routing.module';
-import { CadastroComponent } from './features/users/cadastro/cadastro.component'
+import { TesteModule } from './features/teste/teste.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    NavbarComponent,
-    CadastroComponent
+    NavbarComponent
   ],
   imports: [
     HttpClientModule,
     BrowserModule,
     AppRoutingModule,
-    FormsModule
+    TesteModule
   ],
   providers: [],
   bootstrap: [AppComponent]
