+++ frontend/src/app/features/teste/pages/cadastro/cadastro.component.html
@@ -13,39 +13,39 @@
             <div class="row">
                 <div class="input-field col s5 form-group">
                     <label for="nome">Nome*</label>
-                    <input id="nome" type="text" class="form-control" [(ngModel)]="user.nome">
+                    <input id="nome" type="text" class="form-control" name="nome" [(ngModel)]="user.nome">
                 </div>
                 <div class="input-field col s5 form-group">
                     <label for="sobrenome">Sobrenome*</label>
-                    <input id="sobrenome" type="text" class="form-control" [(ngModel)]="user.sobrenome">
+                    <input id="sobrenome" type="text" class="form-control" name="sobrenome" [(ngModel)]="user.sobrenome">
                 </div>
             </div>
             <div class="row">
                 <div class="input-field col s10 form-group">
                     <label for="cpf">CPF*</label>
-                    <input id="cpf" type="text" class="form-control" [(ngModel)]="user.cpf">
+                    <input id="cpf" type="text" class="form-control" name="cpf" [(ngModel)]="user.cpf">
                 </div>
             </div>
             <div class="row">
                 <div class="input-field col s5 form-group">
                     <label for="telefone">Telefone</label>
-                    <input id="telefone" type="number" class="form-control" [(ngModel)]="user.telefone">
+                    <input id="telefone" type="number" class="form-control" name="telefone" [(ngModel)]="user.telefone">
                 </div>
                 <div class="input-field col s5 form-group">
                     <label for="dataNascimento">Data de Nascimento</label>
-                    <input id="dataNascimento" type="text" class="form-control" [(ngModel)]="user.dataNascimento">
+                    <input id="dataNascimento" type="text" class="form-control" name="dataNascimento" [(ngModel)]="user.dataNascimento">
                 </div>
             </div>
             <div class="row">
                 <div class="input-field col s10 form-group">
                     <label for="email">Email*</label>
-                    <input id="email" type="email" class="form-control" [(ngModel)]="user.email">
+                    <input id="email" type="email" class="form-control" name="email" [(ngModel)]="user.email">
                 </div>
             </div>
             <div class="row">
                 <div class="input-field col s10 form-group">
                     <label for="senha">Senha*</label>
-                    <input id="senha" type="password" class="form-control" [(ngModel)]="user.senha">
+                    <input id="senha" type="password" class="form-control" name="senha" [(ngModel)]="user.senha">
                 </div>
             </div>
             <button class="btn waves-effect waves-light offset-s2" type="submit" value="enviar">Enviar
