+++ frontend/src/proxy.conf.json
@@ -1,0 +1,6 @@
+{
+    "/api": {
+        "target": "http://localhost:8000",
+        "secure": false
+    }
+}
